scheduler/pyproject.toml
[tool.poetry]
name = "scheduler"
version = "0.1.0"
description = ""
authors = ["Cyril Duval <cyril.duval@diabolocom.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pyyaml = "^6.0.2"
apscheduler = "^3.10.4"
sentry-sdk = "^2.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
